06 - Trimestre/IQO_Project/IQO/requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-import-export==3.3.4
django-jazzmin==2.6.0
et-xmlfile==1.1.0
MarkupPy==1.14
mysql-connector-python==8.2.0
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
protobuf==4.21.12
PyMySQL==1.1.0
python-slugify==8.0.1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
